hw03/requirements.txt
streamlit
face-recognition
numpy
Pillow